actions/requirements-actions.txt
numpy==1.21.0
pydantic==1.8.2
pymongo==4.0.1
pytz==2021.3
scikit_learn==1.0.2
SQLAlchemy==1.4.27
datetime
requests
pymongo[srv]
python-dotenv==0.19.2